+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## v52.x-v53 (In Progress)
 - Music: Multiple tracks by Aurilliux added.
 - Music: Christmas trackes removed from non-Christmas seeds; unsafe tracks removed from "t" seeds.
+- PC sprites: Added Andro, Maz, Nero, and Dirk.
 
 ## v52 (2023-06-15)
 - Holiday: Fourth of July option added.